Add NonBlockingConcurrentQueue, NonBlockingExchanger and NonBlockingFutureHolder implementations in JAVA and C#.

+++ C#/NonBlockingFutureHolderSynchronizer/NonBlockingFutureHolderSynchronizer/IFutureHolder.cs
@@ -9,7 +9,7 @@ namespace NonBlockingFutureHolderSynchronizer
     interface IFutureHolder<T>
     {
         void SetValue(T value);
-        T GetValue(long timeout);
+        T GetValue(int timeout);
         bool IsValueAvailable();
     }
 }
